+++ src/capytaine/bodies/bodies.py
@@ -531,15 +531,6 @@ class FloatingBody(_FloatingBodyHydrostaticsMixin, AbstractBody):
             yield "center_of_mass", tuple(self.center_of_mass)
         yield "name", self.name
 
-    def show(self, *args, **kwargs):
-        return self.mesh.show(*args, **kwargs)
-
-    def show_pyvista(self, *args, **kwargs):
-        return self.mesh.show_pyvista(*args, **kwargs)
-
-    def show_matplotlib(self, *args, **kwargs):
-        return self.mesh.show_matplotlib(*args, **kwargs)
-
     def animate(self, motion, *args, **kwargs):
         raise NotImplementedError("Not yet re-implemented in version 3")